requirements.txt
alpaca-py==0.7.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
msgpack==1.0.4
numpy==1.24.1
pandas==1.5.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
sseclient-py==1.7.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
websockets==10.4
